feature: 支持跨项目引用构建机集群（环境）issue #4582
