+++ src/frontend/locale/environment/en-US.json
@@ -2,6 +2,9 @@
     "environment": "Pools",
     "node": "Agent",
     "project": "Project",
+    "operateUser": "Operator",
+    "operateTime": "Opertate Time",
+    "addProject": "Add Project",
     "environmentManage": "Pools",
     "environmentList": "Pools list",
     "createEnvrionment": "Create pool",
@@ -101,6 +104,10 @@
     "mirrorTagRule": "Lower-case letters, numbers, underline and dots only",
     "selectModel": "Select model",
     "confirmInfo": "Confirm information",
+    "setting": "Setting",
+    "removeShareEnvProjectSuc": "project {0} removed",
+    "shareEnvSuc": "Share Env Successful",
+    "removeShareProjectConfirm": "{0} will be removed",
     "envInfo": {
         "createVariable": "Add",
         "name": "Name",
