+++ src/frontend/locale/environment/zh-CN.json
@@ -1,7 +1,10 @@
 {
     "environment": "环境",
     "node": "节点",
-    "project": "项目",
+    "project": "项目",    
+    "operateUser": "操作人",
+    "operateTime": "操作时间",
+    "addProject": "添加项目",
     "environmentManage": "环境管理",
     "environmentList": "环境列表",
     "createEnvrionment": "新增环境",
@@ -101,6 +104,10 @@
     "mirrorTagRule": "只允许包含大小写英文，数字，下划线和点",
     "selectModel": "选择机型",
     "confirmInfo": "确认信息",
+    "setting": "设置",
+    "removeShareEnvProjectSuc": "项目 {0} 移除成功",
+    "shareEnvSuc": "共享环境成功",
+    "removeShareProjectConfirm": "是否移除项目 {0}",
     "envInfo": {
         "createVariable": "新增变量",
         "name": "名称",
